+++ include/firesteel/renderer.hpp
@@ -34,7 +34,9 @@ namespace Firesteel {
         virtual void setDepthTestType(const DepthTestType& tVal) { }
 
         virtual void setViewportSize(const int& tX, const int& tY) { }
-        virtual void setViewportSize(const glm::vec2& tSize) { }
+        void setViewportSize(const glm::vec2& tSize) {
+            setViewportSize(static_cast<int>(tSize.x), static_cast<int>(tSize.y));
+        }
 
         virtual void imguiInitialize(GLFWwindow* tWin) { }
         virtual void imguiNewFrame() {}
