+++ include/firesteel/utils/all.hpp
@@ -1,0 +1,12 @@
+#ifndef FS_UTILS_H
+#define FS_UTILS_H
+
+#include "color.hpp"
+#include "date_time.hpp"
+#include "delta_time.hpp"
+#include "math.hpp"
+#include "os.hpp"
+#include "random.hpp"
+#include "string.hpp"
+
+#endif // !FS_UTILS_H
